Define stateful Accordion
This is the first PF4 component defined that is _stateful_.

What I've done might be a horrible mistake or prove difficult to use. I
have tried to encapsulate implementation to allow me to the approach.

In PatternFly terms, borrowing from React, Accordion is "controlled", or
driven internally by it's events and is relying _state_. Unlike the
`PF4.Navigation`, the `PF4.Accordion` has a `Msg`, an opaque `State`, &
an implementation of `update` that must be called by the parent Elm
application. This also means that `PF4.Accordion` needs a way for its
`Msg` to be lifted or _mapper_ into terms of `msg`, the parent Elm's
`msg` type.

This is not a complete visual appearance mapping to PatternFly's design
system yet. There are a couple of things missing (like bordered
rendering, and influencing the text size - I still don't know what in
the world **FIXED** means in the example, see link below).

https://www.patternfly.org/v4/components/accordion