Reversion to 1.0.0
Note: this package is really in a 0.0.1 state, please treat it as such.

When attempting to `elm publish` I got the message below:

```
I will now verify that everything is in order...

  + Found README.md
  + Found LICENSE
  + Verified documentation
  X Version 0.0.1 is not correct!

-- INVALID VERSION -----------------------------------------------------

I cannot publish 0.0.1 as the initial version.

Change it to 1.0.0 which is the initial version for all Elm packages.
```

I do appreciate making the versioning explicitly; **AND** I wish there
was a way to indicate this package is _totally_ "alpha".

I will keep the following in mind:

```
  - Versions are incremented based on how the API changes:

        PATCH = the API is the same, no risk of breaking code
        MINOR = values have been added, existing values are unchanged
        MAJOR = existing values have been changed or removed
```

As I add new PF4 components that are not in 1.0.0, I will be using those
MINOR version bumps whenever I can.