Attach onClick to parent container (not child)
Wow! This was an annoying one for me as I like to click along the center
of the navItems and with the labels (child of container) being "ragged
right", I was hitting situations where a click didn't change the page.

Why? Well, at first I think it was BLACK MAGIC (and the universe hating
me). But it is because the `itemMarkup` was associating the onClick
with the child (a lowercase l, label). It was possible to click on the
Element.row without it receiving that click.

Much Embarrassment has ensued. And, I'm delighted this is fix.